Rename AppRunner to Swift.Bindings.Framework.Tests
