+++ src/Swift.Bindings/tests/FrameworkTests/Platforms/MacCatalyst/Entitlements.plist
