+++ src/Swift.Bindings/tests/FrameworkTests/Platforms/MacCatalyst/Info.plist
