+++ src/Swift.Bindings/tests/FrameworkTests/Platforms/iOS/Info.plist
