+++ src/Swift.Bindings/tests/FrameworkTests/Platforms/iOS/Resources/PrivacyInfo.xcprivacy
