+++ src/Swift.Bindings/tests/FrameworkTests/Resources/Fonts/FluentUI.cs
