+++ src/Swift.Bindings/tests/FrameworkTests/Resources/Fonts/OpenSans-Regular.ttf
