+++ src/Swift.Bindings/tests/FrameworkTests/Resources/Fonts/OpenSans-Semibold.ttf
