+++ src/Swift.Bindings/tests/FrameworkTests/Resources/Images/dotnet_bot.png
