+++ src/Swift.Bindings/tests/FrameworkTests/Resources/Raw/AboutAssets.txt
