+++ src/Swift.Bindings/tests/FrameworkTests/Resources/Styles/Colors.xaml
