+++ src/Swift.Bindings/tests/FrameworkTests/Resources/Styles/Styles.xaml
