+++ src/Swift.Bindings/tests/FrameworkTests/Swift.Bindings.Framework.Tests.csproj
@@ -3,7 +3,7 @@
 	<PropertyGroup>
 		<TargetFramework>net9.0-maccatalyst</TargetFramework>
 		<OutputType>Exe</OutputType>
-		<RootNamespace>RunnerApp</RootNamespace>
+		<RootNamespace>BindingsGeneration.FrameworkTests</RootNamespace>
 		<UseMaui>true</UseMaui>
 		<SingleProject>true</SingleProject>
 		<ImplicitUsings>enable</ImplicitUsings>
@@ -11,10 +11,10 @@
     <IsTestProject>false</IsTestProject>
 
 		<!-- Display name -->
-		<ApplicationTitle>RunnerApp</ApplicationTitle>
+		<ApplicationTitle>SwiftBindings</ApplicationTitle>
 
 		<!-- App Identifier -->
-		<ApplicationId>com.companyname.runnerapp</ApplicationId>
+		<ApplicationId>com.companyname.swiftbindings</ApplicationId>
 
 		<!-- Versions -->
 		<ApplicationDisplayVersion>1.0</ApplicationDisplayVersion>
